+++ src/main/java/com/silverminer/shrines/registries/RandomVariationMaterialRegistry.java
@@ -8,24 +8,16 @@
 
 package com.silverminer.shrines.registries;
 
-import com.mojang.datafixers.util.Pair;
 import com.silverminer.shrines.Shrines;
 import com.silverminer.shrines.random_variation.RandomVariationMaterial;
-import com.silverminer.shrines.registry.ModRegistrar;
-import com.silverminer.shrines.registry.ModRegistryObject;
-import net.minecraft.world.level.block.Blocks;
+import net.minecraftforge.registries.DeferredRegister;
+import net.minecraftforge.registries.IForgeRegistry;
+import net.minecraftforge.registries.RegistryBuilder;
 
-import java.util.List;
+import java.util.function.Supplier;
 
 public class RandomVariationMaterialRegistry {
-   public static final ModRegistrar<RandomVariationMaterial> REGISTRY = ModRegistrar.create(Shrines.MODID, RandomVariationMaterial.REGISTRY);
-   public static final ModRegistryObject<RandomVariationMaterial> DARK_OAK = REGISTRY.register("dark_oak", () -> new RandomVariationMaterial(List.of(
-         Pair.of(Blocks.DARK_OAK_PLANKS, Shrines.location("planks"))
-   )));
-   public static final ModRegistryObject<RandomVariationMaterial> BIRCH = REGISTRY.register("birch", () -> new RandomVariationMaterial(List.of(
-         Pair.of(Blocks.BIRCH_PLANKS, Shrines.location("planks"))
-   )));
-   public static final ModRegistryObject<RandomVariationMaterial> OAK = REGISTRY.register("oak", () -> new RandomVariationMaterial(List.of(
-         Pair.of(Blocks.OAK_PLANKS, Shrines.location("planks"))
-   )));
+   public static final DeferredRegister<RandomVariationMaterial> REGISTRY = DeferredRegister.create(RandomVariationMaterial.REGISTRY, Shrines.MODID);
+   public static final Supplier<IForgeRegistry<RandomVariationMaterial>> FORGE_REGISTRY_SUPPLIER = REGISTRY.makeRegistry(RandomVariationMaterial.class,
+         () -> new RegistryBuilder<RandomVariationMaterial>().dataPackRegistry(RandomVariationMaterial.DIRECT_CODEC));
 }